Use byte as index for known type names
Uses a byte instead of an int to index known type names. We'll need to revisit this if the number of known types grows beyond 256.

To make serialization a bit easier, I've added methods for reading/writing bytes to `JsonDataReader` and `JsonDataWriter`.

+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Serialization/MessagePack/Formatters/TagHelpers/TypeNameObjectFormatter.cs
@@ -37,7 +37,7 @@ internal sealed class TypeNameObjectFormatter : ValueFormatter<TypeNameObject>
         switch (typeNameKind)
         {
             case TypeNameKind.Index:
-                var index = reader.ReadInt32();
+                var index = reader.ReadByte();
                 return new(index);
             case TypeNameKind.String:
                 var fullName = CachedStringFormatter.Instance.Deserialize(ref reader, options).AssumeNotNull();
@@ -58,7 +58,7 @@ internal sealed class TypeNameObjectFormatter : ValueFormatter<TypeNameObject>
 
         writer.WriteArrayHeader(PropertyCount);
 
-        if (value.Index is int index)
+        if (value.Index is byte index)
         {
             writer.Write((byte)TypeNameKind.Index);
             writer.Write(index);
